Calendar Matching Session

+++ demo/src/main/java/User/DTO/Response/TutorProfileResponse.java
@@ -28,5 +28,5 @@ public class TutorProfileResponse {
 
     private VerificationStatus verificationStatus;
     private Double averageRating;
-    private Integer trustScore;
+    private Double trustScore;
 }
